Add InMemory email sender
