+++ tests/GeekLearning.Email.Integration.Test/project.json
@@ -42,7 +42,7 @@
         }
       }
     },
-    "net451": {
+    "net46": {
       "frameworkAssemblies": {
       },
       "dependencies": {
